LayoutTests/imported/w3c/web-platform-tests/screen-orientation/fullscreen-interactions.html

<!DOCTYPE html>
<meta charset="utf-8" />
<meta viewport="width=device-width, initial-scale=1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body></body>
<script type="module">
  import {
    getOppositeOrientation,
    attachIframe,
    makeCleanup,
  } from "./resources/orientation-utils.js";

  promise_test(async (t) => {
    t.add_cleanup(makeCleanup());
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    const currentOrientation = screen.orientation.type;
    await screen.orientation.lock(
      getOppositeOrientation()
    );
  }, "fullscreen and orientation support");

  promise_test(async (t) => {
    const iframe = await attachIframe();
    t.add_cleanup(makeCleanup(iframe));
    const iframeWindow = iframe.contentWindow;
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    const currentOrientation = window.screen.orientation.type;
    const lockPromise = iframeWindow.screen.orientation.lock(
      getOppositeOrientation()
    );
    await document.exitFullscreen();
    const frameDOMException = iframe.contentWindow.DOMException;
    await promise_rejects_dom(t, "AbortError", frameDOMException, lockPromise);
  }, "Fully unlocking the screen orientation causes a pending lock in a nested browsing context to be aborted");

  promise_test(async (t) => {
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    const { orientation } = screen;
    const oppositeOrientation = getOppositeOrientation();
    const lockPromise = orientation.lock(oppositeOrientation);
    const fsExitPromise = document.exitFullscreen();
    await promise_rejects_dom(t, "AbortError", lockPromise);
    await fsExitPromise;
  }, "Exiting fullscreen causes a pending lock to be aborted");

  promise_test(async (t) => {
    const iframe = await attachIframe();
    t.add_cleanup(makeCleanup(iframe));
    const iframeWindow = iframe.contentWindow;
    await test_driver.bless("request full screen", null, iframeWindow);
    await iframe.contentDocument.documentElement.requestFullscreen();
    const lockPromise = iframeWindow.screen.orientation.lock(
      getOppositeOrientation()
    );
    const fsExitPromise = iframe.contentDocument.exitFullscreen();
    await promise_rejects_dom(
      t,
      "AbortError",
      iframeWindow.DOMException,
      lockPromise
    );
    await fsExitPromise;
  }, "Exiting fullscreen from an iframe causes a pending lock to be aborted");
</script>
